chore: configure tailwind setting and rename file
