+++ src/components/recordCard/Daily/index.tsx
@@ -1,23 +1,9 @@
-import { IconX } from "@tabler/icons-react";
 import ToggleList from "@/components/ToggleList";
+import CardWrapper from "../Wrapper";
 
-const Daliy = () => {
+const Daily = () => {
   return (
-    <section className="flex flex-col gap-6 p-8 w-[416px] border border-stroke rounded-[30px]">
-      <div>
-        <div className="flex justify-between items-center">
-          <h2 className="text-2xl font-bold">
-            <span className="inline-block mr-4 w-[11px] h-[11px] rounded-full bg-[#969AFF]"></span>
-            新增日常紀錄
-          </h2>
-          <button type="button">
-            <IconX size={32} stroke={2} />
-          </button>
-        </div>
-        <h3 className="mt-1 text-note">
-          可讓醫生參考的生理紀錄及在家可處理的照護。
-        </h3>
-      </div>
+    <CardWrapper>
       <form className="flex flex-col gap-4">
         <ToggleList title="一般">
           <ul className="flex flex-col gap-4 mt-2">
@@ -105,7 +91,7 @@ const Daliy = () => {
           登入
         </button>
       </form>
-    </section>
+    </CardWrapper>
   );
 };
-export default Daliy;
+export default Daily;
