+++ tailwind.config.ts
@@ -22,6 +22,9 @@ const config: Config = {
         stroke: "#EAEAEA",
         error: "#F23030",
         link: "#0057FF",
+        medical: "FF6D80",
+        daily: "969AFF",
+        moment: "FFA959",
       },
     },
   },
